requirements.txt
openai
python-dotenv
pyppeteer
requests
